Refactor
- Added environment variables: SECRET_KEY, DATABASE_URL, ALGORITHM
- Moved functions: hash_password(password:), verify_password(plain_password:, hashed_password:)
- Added blank lines (PEP 8)
