+++ sennight-fastapi/app/crud.py
@@ -1,21 +1,9 @@
 from sqlalchemy.orm import Session
-import bcrypt
-from . import models, schemas
+from . import models, schemas, security
 
-# Hash a password using bcrypt
-def hash_password(password: str) -> str:
-    pwd_bytes = password.encode('utf-8')
-    salt = bcrypt.gensalt()
-    hashed_password = bcrypt.hashpw(password=pwd_bytes, salt=salt)
-    return hashed_password.decode('utf-8')
-
-# Check if the provided password matches the stored password (hashed)
-def verify_password(plain_password: str, hashed_password: str) -> bool:
-    password_byte_enc = plain_password.encode('utf-8')
-    return bcrypt.checkpw(password = password_byte_enc , hashed_password = hashed_password.encode('utf-8'))
 
 def create_user(db: Session, user: schemas.UserCreate):
-    hashed_password = hash_password(user.password)
+    hashed_password = security.hash_password(user.password)
     db_user = models.User(
         email=user.email,
         name=user.name,
@@ -29,10 +17,11 @@ def create_user(db: Session, user: schemas.UserCreate):
     db.refresh(db_user)
     return db_user
 
+
 def authenticate_user(db: Session, email: str, password: str):
     user = db.query(models.User).filter(models.User.email == email).first()
     if not user:
         return False
-    if not verify_password(password, user.hashed_password):
+    if not security.verify_password(password, user.hashed_password):
         return False
     return user
