+++ sennight-fastapi/app/models.py
@@ -1,6 +1,7 @@
 from sqlalchemy import Column, Integer, String, Date, Float, DateTime, func
 from .database import Base
 
+
 class User(Base):
     __tablename__ = "users"
     id = Column(Integer, primary_key=True, autoincrement=True)
